Rename Plugin to ServiceLayerPlugin, keep Plugin as deprecated BC alias
